requirements/testing.txt
# Package Requirements for Testing
beautifulsoup4 >= 4.3.2, < 5.0.0
coverage == 4.0a5
factory_boy == 2.5.2
flake8 == 2.4.0
pep8-naming == 0.2.2
pretend >= 1.0.7, < 2.0.0
pysqlite >= 2.6.3, < 3.0.0
pytest-django == 2.8.0
